# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 学习状态智能助手 API（板块一 + 板块二 MVP）
  version: 1.8.0
  description: |
    中学学情导学与状态评估 API。

    来源：`docs/api-design-unified.md`（唯一生效版本），对应 `PRD-学习状态智能助手.md`。
    本文件严格按该设计文档生成，未引入文档之外的字段。

    主流程：创建目标 ① → 生成计划 ② → 提交学习记录 ③ → 计算状态 ④ → 返回建议 ⑤（外加周期复盘与反馈回路）。

    约定要点：
    - 所有资源以当前 token 用户为作用域，不设 userId 路径参数，不存在跨用户读取接口。
    - 写接口可选携带 `Idempotency-Key` 头，24 小时内重复键直接返回首次结果。
    - 建议与复盘为异步生成：POST 返回 202 + `generation.status = pending`，前端轮询 GET 获取结果。
    - 监护人授权失效时所有写接口返回 403 `GUARDIAN_AUTHORIZATION_EXPIRED`，读接口不受影响。

    阶段口径：pilot。本文档为前后端与 QA 的唯一契约来源。项目当前处于 pilot 阶段（小范围试用验证，非 beta/正式发布），由学生团队维持。板块三契约已于 M1 转正（口径见 docs/module3-m0-decision-proposal.md v1.7 §4.11）。

    **v1.6.0（重构 M0 · 契约冻结）**——本次为重构第一阶段的契约增量，未改动任何既有字段名：
    - **稳定用户 ID 落地（D59）**：`User.userId` 语义明确为稳定内部主键（`u_` 前缀），与登录凭证
      解耦；`AuthMeResponse.user` 增补可选 `userId`（`email` 保留，前端展示仍用它）。
    - **新增 M0 实体 schema**：Exam / Collection / CollectionItem / UsageLedgerEntry / InviteCode /
      ViolationLog / ErrorReport / Medal / UserProfileEntry / TopicSummary / Explanation /
      ChatSession / ChatRawMessage。⚠️ 这些 schema **尚未挂到任何 path**，接口由各板块按里程碑补齐；
      字段名以本节为准，各板块不得另起。
    - **题本升格（D48）**：新增 `ErrorCause` / `ErrorIntent` 两个正交维度，`ErrorRecord` 系列扩展
      `errorCause` / `intent` / `sourceExamId` 三个可选字段（`errorType` 保留为自由文本，向后兼容）。
    - **目标（D6 / D49）**：`Goal` / `GoalSummary` / `GoalCreate` / `GoalUpdate` 增补
      `parentGoalId`（父子树）、`examId` + `targetScore`（引用考试）。`status` 的 enum 未改动。
...

...
servers:
...
security:
  - sessionCookie: []
...
paths:
...
  /me/usage:
    get:
      tags: [pilot 运营与治理]
      operationId: getMyUsage
      summary: 只读用量查询（按月返回当前用户的模型数值成本流水与合计）
      description: |
        用量页数据源（D38/#9）。**只读，无任何支付/充值语义**——支付整体推迟到 pilot 结束后。
        返回 usage_ledger 流水（只存数值，不含任何提示词或输出内容）与区间合计；
        month 缺省为当前月。与刻意去身份化的 AICallLog 分两套，不合并。
      parameters:
        - name: month
          in: query
          required: false
          description: 查询月份 YYYY-MM；缺省为当前月
          schema:
            type: string
            pattern: '^\d{4}-\d{2}$'
      responses:
        '200':
          description: 区间内用量流水与合计
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageLedgerList'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /error-reports:
    post:
      tags: [pilot 运营与治理]
      operationId: createErrorReport
      summary: 用户报错（设置常驻入口 messageId=null；消息级按钮带 messageId 与上下文）
      description: |
        #46 两处入口共用的后端。消息级报错必须带上下文（哪条消息、什么意图），
        context 只放结构化快照，**不含对话原文流水**（与 D45 原文短期留存口径分开）。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ErrorReportCreate'
      responses:
        '201':
          description: 报错已受理
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorReport'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /me/violations:
    get:
      tags: [pilot 运营与治理]
      operationId: getMyViolations
      summary: 本人违规处置留痕（warn / temp_ban / perm_ban，按时间倒序）
      description: |
        #43 分级处置（1 次警告 → 3 次临时封禁 → 永久）必须留痕；本人可查自己的处置记录。
        reason 只含命中类型，不含用户原文。
      responses:
        '200':
          description: 处置留痕列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ViolationLogList'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /me/medals:
    get:
      tags: [pilot 运营与治理]
      operationId: getMyMedals
      summary: 本人已获得奖章（#49 最小版：里程碑见 MedalMilestone，未获得的不返回）
      description: |
        虚拟奖章最小版，只列已获得里程碑。**不做积分商城 / 排行榜**
        （排行榜与 §4.5「社区不做社交」冲突）。
      responses:
        '200':
          description: 已获得奖章列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedalList'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /analytics/events:
    post:
      tags: [pilot 运营与治理]
      operationId: createAnalyticsEvent
      summary: 结构化埋点事件上报（D39 三块；不含对话原文流水，payload 入库前脱敏）
      description: |
        D39 三块埋点统一上报口：chat_interaction（Chat 交互）/ ai_quality（AI 质量反馈）/
        profile_trace（画像 trace）。**与对话原文的口径不可混同**：本接口只收结构化事件
        （意图判定结果、卡片事件、锚定结果、提取事件…），不含对话原文流水；
        payload 入库前经 privacy_filter 脱敏，身份只走服务端会话，不进 payload。
        id / createdAt 由服务端生成；响应返回脱敏后实际落库的内容。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyticsEventCreate'
      responses:
        '201':
          description: 事件已落库（返回脱敏后的实际存储内容）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsEvent'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sid
      description: |
        HttpOnly Session Cookie（名为 sid），登录成功后由服务端 Set-Cookie 下发，
        有效期 7 天（见 mock-server SESSION_TTL_MS）。浏览器自动携带，前端无需手动设置。
        鉴权失效时业务接口返回 401 UNAUTHENTICATED。
        （历史说明：早期契约曾假设上游 Bearer token 且不含登录/注册接口，
        实际项目改为自带 /auth/* 会话体系，本契约已据此收录，见「鉴权会话」标签。）
...
  responses:
    ValidationFailed:
      description: 参数缺失或取值越界
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: VALIDATION_FAILED
              message: 自评专注度必须为 1-5 的整数
              field: selfReport.focus
    Unauthenticated:
      description: token 缺失或过期
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: UNAUTHENTICATED
              message: 登录状态已失效，请重新登录
...
    InternalServerError:
      description: 服务端错误。对齐 api-design-unified.md 0.2 节错误码表中的 500 INTERNAL_ERROR。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: INTERNAL_ERROR
              message: 服务开小差了，请稍后再试
...
  schemas:
...
    Error:
      type: object
      description: 统一错误格式
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: 错误码
            message:
              type: string
              description: 可展示的错误说明
            field:
              type: string
              description: 校验失败的字段路径，仅参数校验错误时出现
...
    UsageFeatureTier:
      type: string
      description: >-
        功能档位（分档定价依据，D38）。对应目标态 §7 的三类 AI + 上传解析：
        chat=对话式（闲聊 / 搜题 / 讲知识点 / 查；用户主动发起、有感）；
        embedded=嵌入式（建议 / 复盘 / 口语转译 / 画像提取 / 推荐卡；系统自动、无感，**不进 credits**）；
        advanced=高级（L3 富交互仿真 / 多模态搜题；稀有且贵，高级月卡或 credits 单次）；
        multimodal=文件解析（上传的图片 / PDF / PPT 理解，F 板块）。
        ⚠️ 新增档位属契约变更，需同步本枚举与定价口径。
      enum: [chat, embedded, advanced, multimodal]
    ReasoningTier:
      type: string
      description: >-
        推理等级（#33：快速 / 标准 / 深度），与 credits 扣费档位绑定。
        随手问浮窗跟随主对话档位，浮窗内不单独给选择。
      enum: [quick, standard, deep]
    UsageLedgerEntry:
      type: object
      description: >-
        按用户记录模型数值成本。**只存数值，不存任何提示词或输出内容**；
        与刻意去身份化的 AICallLog 分两套，不合并（绝不给 AICallLog 补身份字段）。
      required: [id, featureTier, model, tokensIn, tokensOut, cost, createdAt]
      properties:
        id: { type: string }
        featureTier:
          $ref: '#/components/schemas/UsageFeatureTier'
        reasoningTier:
          allOf:
            - $ref: '#/components/schemas/ReasoningTier'
          nullable: true
        model: { type: string }
        tokensIn: { type: integer }
        tokensOut: { type: integer }
        cost: { type: number, description: 数值成本（内部计价，非对用户计费） }
        createdAt:
          type: string
          format: date-time
    UsageLedgerList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UsageLedgerEntry'
        totalCost:
          type: number
          description: 区间内数值成本合计
        totalTokensIn: { type: integer }
        totalTokensOut: { type: integer }

    # ---------- A 板块：邀请码（#50） ----------
...
    ViolationAction:
      type: string
      description: 分级处置动作（1 次警告 → 3 次临时封禁 → 永久封禁），必须留痕
      enum: [warn, temp_ban, perm_ban]
    ViolationLog:
      type: object
      required: [id, level, action, reason, createdAt]
      properties:
        id: { type: string }
        level:
          type: integer
          description: 累计违规等级（1 起）
        action:
          $ref: '#/components/schemas/ViolationAction'
        reason: { type: string, description: 处置原因（命中类型，不含原文） }
        createdAt: { type: string, format: date-time }

    # ---------- G 板块：报错入口（#46） ----------
    ErrorReport:
      type: object
      description: 用户报错。两处入口：设置常驻 + 每条模型输出旁的消息级按钮；必须带上下文。
      required: [id, createdAt]
      properties:
        id: { type: string }
        messageId:
          type: string
          nullable: true
          description: 消息级报错时的消息 ID（设置常驻入口为 null）
        intent:
          type: string
          nullable: true
          description: 该条消息的意图判定结果（便于定位问题）
        description:
          type: string
          description: 用户描述的问题（回读用，与 ErrorReportCreate.description 对应）
        context:
          type: object
          nullable: true
          description: 上下文快照（消息角色 / 结构化体等，**不含对话原文流水**）
        createdAt: { type: string, format: date-time }
    ErrorReportCreate:
      type: object
      required: [description]
      properties:
        messageId: { type: string, nullable: true }
        intent: { type: string, nullable: true }
        description:
          type: string
          maxLength: 2000
          description: 用户描述的问题
        context:
          type: object
          nullable: true

    # ---------- G 板块：奖章（#49） ----------
    MedalMilestone:
      type: string
      description: pilot 只做最小版里程碑；不做积分商城 / 排行榜（与「社区不做社交」冲突）
      enum: [first_record, streak_7_days, first_summary, first_goal_achieved, first_topic_book]
    Medal:
      type: object
      required: [medalId, milestone, awardedAt]
      properties:
        medalId: { type: string }
        milestone:
          $ref: '#/components/schemas/MedalMilestone'
        awardedAt: { type: string, format: date-time }
    MedalList:
      type: object
      description: G 板块增量（2026-09-25）：本人已获得奖章列表（#49 最小版）
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Medal'
    ViolationLogList:
      type: object
      description: G 板块增量（2026-09-25）：本人违规处置留痕列表（#43）
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ViolationLog'
    AnalyticsEventCreate:
      type: object
      description: >-
        G 板块增量（2026-09-25）：埋点上报体（D39）。id / createdAt 由服务端生成；
        payload 只收结构化事件体，**不含对话原文流水**，入库前经 privacy_filter 脱敏。
      required: [category, eventType]
      properties:
        category:
          $ref: '#/components/schemas/AnalyticsCategory'
        eventType:
          type: string
          maxLength: 64
          description: >-
            事件名，如 message_sent / card_shown / card_clicked / intent_corrected /
            stack_pushed / stack_sunk / suggestion_rated / profile_extracted
        sessionId: { type: string, nullable: true }
        payload:
          type: object
          nullable: true
          description: 结构化事件体（不含对话原文流水；脱敏后入库）
        occurredAt:
          type: string
          format: date-time
          nullable: true
          description: 事件业务发生时间；缺省为服务端接收时间

    # ---------- B 板块：画像（D50） ----------
...
    AnalyticsCategory:
      type: string
      description: >-
        D39 的三块埋点：chat_interaction=Chat 交互（消息/卡片/意图纠正/上下文栈/引导式搜题交互）；
        ai_quality=AI 质量反馈（读完率/采纳率/有用性/纠错）；profile_trace=画像 trace（提取与改删事件）。
      enum: [chat_interaction, ai_quality, profile_trace]
    AnalyticsEvent:
      type: object
      description: >-
        结构化埋点事件。⚠️ **与对话原文的口径不可混同**（§3.8.3 / §8 反复强调）：
        本表存的是**结构化事件**（意图判定结果、卡片事件、锚定结果、提取事件…），
        **不含对话原文流水**——原文属 `chat_raw_messages` 的短期安全留存（产品不可见、到期物理删除）。
        `payload` 入库前走 privacy_filter 脱敏。
      required: [id, category, eventType, occurredAt]
      properties:
        id: { type: string }
        category:
          $ref: '#/components/schemas/AnalyticsCategory'
        eventType:
          type: string
          description: >-
            事件名，如 message_sent / card_shown / card_clicked / intent_corrected /
            stack_pushed / stack_sunk / suggestion_rated / profile_extracted
        sessionId: { type: string, nullable: true }
        payload:
          type: object
          nullable: true
          description: 结构化事件体（**不含对话原文流水**；脱敏后入库）
        occurredAt:
          type: string
          format: date-time
          description: 事件业务发生时间（与落库时间区分）
        createdAt: { type: string, format: date-time }

    # ---------- D 板块：搜题归档（§3.8.4 / D24 / D52） ----------
...